refactor masterlist
